Refactor history cleanup and improve logging
Changed DeleteOldHistory to return deleted records and updated cleanup logic to remove corresponding image files. Enhanced logging in MonitorService and updated namespaces to reflect refactoring. Commented out old orphaned file cleanup logic for clarity.
